Split routes and FastAPI app + make custom class overriding FastAPI class
